refactor: rename parser.ts to index.ts
